# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/wallet/{userId}": {
      "get": {
        "operationId": "WalletController_get",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Wallet"
        ]
      }
    },
    "/api/wallet/{userId}/ledger": {
      "get": {
        "operationId": "WalletController_ledger",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Wallet"
        ]
      }
    },
    "/api/wallet/{userId}/adjust": {
      "post": {
        "operationId": "WalletController_adjust",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Wallet"
        ]
      }
    },
...
  },
  "info": {
    "title": "IPEasy Platform API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "apikey": {
        "type": "apiKey",
        "in": "header",
        "name": "apikey"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}